### 2020-07-12 #2477 @whedon
[PRE REVIEW]: DiffCapAnalyzer: A Python Package for Quantitative Analysis of Total Differential Capacity Data
**Submitting author:** @nicolet5 (<a href="http://orcid.org/0000-0003-3411-7373">Nicole  Thompson</a>)
**Repository:** <a href="https://github.com/nicolet5/DiffCapAnalyzer" target ="_blank">https://github.com/nicolet5/DiffCapAnalyzer</a>
**Version:** v0.1.1
