### 2020-08-29 #2477 @nicolet5
@labarba Thank you for pointing that out! I wasn't previously aware that was a needed section. I have updated the paper to include the _Statement of need_ section.
