### 2020-08-30 #2477 @jgostick
Hello @yangbai90, you have listed yourself as volunteer JOSS reviewer, and have listed 'battery modeling' as an area of expertise.  I was hoping you'd be interested in reviewing this contribution, since it's about batteries...though more from a systems perspective and less about FEM based performance.  Please let me know here in this thread or contact me by the email address associated with my Github account.
