### 2020-09-01 #2624 @jgostick
Hi @WardLT, you asked in the 'pre-review' thread how to get started.  I have added a checklist to the top of this issues under your name...you now just check off the boxes, and give the author(s) of the package any tips on how to improve their code for any of the boxes that are not 'check-worthy'.
