### 2020-09-01 #2477 @jgostick
@arfon something odd happened on this issue when I assigned reviewers.  I assigned @WardLT, then assigned @yangbai90, and whedon did something unexpected (I think) as you can see by his/her comments above.  Is this buggy or did I screw up?
