### 2020-09-01 #2477 @arfon
@jgostick `assign` clobbers all existing assignments, `add` adds a new reviewer so your second comment here: https://github.com/openjournals/joss-reviews/issues/2477#issuecomment-684181713 removed the first reviewer.

You can fix this over in #2624 by:

1. Running `@whedon add @WardLT as reviewer`
2. Editing the issue at the top of the thread to add a second checklist for @WardLT

Are you OK to do this? If not, I can help.
