### 2020-09-22 c0eecfb
updated readme with contributions section and small change to paper
