### 2020-09-22 #2624 @nicolet5
@yangbai90 Hi Yang, thank you for reviewing the package and for your comments! With regards to your questions:

1. Yes the package has autotests. You can run all the tests using `./runTests` from the root directory, or by using pytest  (see the [Running Tests](https://github.com/nicolet5/DiffCapAnalyzer#running-tests) section of the README. Additionally, I have TravisCI set up, which runs the tests every time a commit is pushed.

2. Thank you for pointing this out! I've updated the README with another section to describe how others can contribute: [Have Any Issues? Want to Contribute?](https://github.com/nicolet5/DiffCapAnalyzer#have-any-issues-want-to-contribute)

3. I've updated the paper's "Statement of Need" section to address this point.

Let me know if there are any other issues I need to address!

Best,
Nicole
