### 2020-10-01 #2624 @WardLT
The software and paper look great to me!

I'm actually looking forward to being able to use this package in my own work. So, I'd like to express my personal appreciation to the authors for making this code available!
