### 2020-10-03 #2624 @nicolet5
Thank you everyone! @WardLT I am glad this will be useful to you and I look forward to seeing how you leverage it!

@jgostick  thank you for the suggestion, I've edited the paper by replacing the last paragraph with a better concluding remark and I agree it reads better now!
