+++ docs/paper/paper.md
@@ -61,7 +61,7 @@ In differential capacity curves without any cleaning or smoothing, there is sign
  
 ![Cleaning process on an example differential capacity curve.\label{Figure 1}](images/cleaning_dqdv.png)
 
-Once the data is clean, the software automatically finds peaks in the dQ/dV curves utilizing the PeakUtils Python package [@peakutils], and returns the peak heights and the peak locations, as shown by an example cycle in \autoref{Figure 2}a. These peak heights and locations are then used to inform the model build, which is individualized to each cycle contained in the dataset. The model consists of Pseudo-Voigt distributions positioned at the identified peak locations and a baseline Gaussian distribution that captures the area that is not part of the Psuedo-Voigt distributions. The Pseudo-Voigt distribution described by Equations 2 and 3 is simply the linear combination of a Gaussian and a Lorentzian. This distribution is often used in fitting experimental spectral data due to it being a highly generalizable function able to fit a large variety of peak shapes [@wertheim_butler_west_buchanan_1974].
+Once the data is clean, the software automatically finds peaks in the dQ/dV curves utilizing the PeakUtils Python package [@peakutils], and returns the peak heights and the peak locations, as shown by an example cycle in \autoref{Figure 2}a. These peak heights and locations are then used to inform the model build, which is individualized to each cycle contained in the dataset. The model consists of Pseudo-Voigt distributions positioned at the identified peak locations and a baseline Gaussian distribution that captures the area that is not part of the Pseudo-Voigt distributions. The Pseudo-Voigt distribution described by Equations 2 and 3 is simply the linear combination of a Gaussian and a Lorentzian. This distribution is often used in fitting experimental spectral data due to it being a highly generalizable function able to fit a large variety of peak shapes [@wertheim_butler_west_buchanan_1974].
 
 $$f_v(x,A,\mu,\sigma,\alpha)=\frac{(1- \alpha)A}{\sigma_g \sqrt{2 \pi}}\exp{[-{(x- \mu)}^2/2 {\sigma_g}^2]}+\frac{\alpha A}{\pi}[\frac{\sigma}{{(x-\mu)}^2 + \sigma^2}]\tag*{(2)}$$
 
