### 2020-10-15 #2624 @nicolet5
Thanks @jgostick

I updated the paper with everyone's ORCIDs, made a release of the software `v1.0.1`, and archived the release on Zenodo with doi `10.5281/zenodo.4089118`. Let me know if there's anything else I need to do!
