### 2020-10-15 #2624 @whedon
:wave: @openjournals/joss-eics, this paper is ready to be accepted and published.
