### 2020-10-15 #2624 @danielskatz
👋 @nicolet5 - this basically looks ready to go, but I have some questions about cases in the references.  I feel like Gaussian and Lorentzian should be capitalized, for example.  And lifepo4 doesn't seem correct either.  Can you fix these, by protecting the capital letters with {}s in the bib file, along with anything else similar that I didn't catch?  Then we can proceed to accept this submission
