### 2020-10-18 7ca2db2
capitalization errors in paper.bib
