+++ docs/paper/paper.bib
@@ -4,7 +4,7 @@
         author={Marzocca, Lauren M. and Atwater, Terrill B.}
         }
 @article{torai_nakagomi_yoshitake_yamaguchi_oyama_2016, 
-        title={State-of-health estimation of LiFePO4/graphite batteries based on a model using differential capacity},
+        title={State-of-health estimation of {L}i{F}e{PO}4/graphite batteries based on a model using differential capacity},
         volume={306}, 
         DOI={10.1016/j.jpowsour.2015.11.070}, 
         journal={Journal of Power Sources}, 
@@ -115,7 +115,7 @@ Publisher: Multidisciplinary Digital Publishing Institute},
         journal={Bitbucket}
         }
 @article{wertheim_butler_west_buchanan_1974, 
-        title={Determination of the Gaussian and Lorentzian content of experimental line shapes}, 
+        title={Determination of the {G}aussian and {L}orentzian content of experimental line shapes}, 
         volume={45}, 
         DOI={10.1063/1.1686503},
         number={11}, 
