+++ docs/paper/paper.bib
@@ -4,7 +4,7 @@
         author={Marzocca, Lauren M. and Atwater, Terrill B.}
         }
 @article{torai_nakagomi_yoshitake_yamaguchi_oyama_2016, 
-        title={State-of-health estimation of {L}i{F}e{PO}4/graphite batteries based on a model using differential capacity},
+        title={State-of-health estimation of {L}i{F}e{PO$_4$}/graphite batteries based on a model using differential capacity},
         volume={306}, 
         DOI={10.1016/j.jpowsour.2015.11.070}, 
         journal={Journal of Power Sources}, 
