+++ docs/paper/paper.bib
@@ -100,7 +100,7 @@ Publisher: Multidisciplinary Digital Publishing Institute},
 	file = {ScienceDirect Full Text PDF:C\:\\Users\\Theodore Cohen\\Zotero\\storage\\F8X8BNEP\\He et al. - 2020 - Comparative study of curve determination methods f.pdf:application/pdf;ScienceDirect Snapshot:C\:\\Users\\Theodore Cohen\\Zotero\\storage\\HJ4SDUE7\\S2352152X19317219.html:text/html}
 }
 @article{bloom_jansen_abraham_knuth_jones_battaglia_henriksen_2005, 
-        title={Differential voltage analyses of high-power, lithium-ion cells}, 
+        title={Differential voltage analyses of high-power, lithium-ion cells: 1. {T}echnique and application}, 
         volume={139}, 
         DOI={10.1016/j.jpowsour.2004.07.021}, 
         number={1-2}, 
